pyproject.toml
[project]
name = "para-mcp"
version = "0.1.0"
description = "Para MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.4.0",
    "httpx[http2]>=0.28.1",
    "mcp[cli]>=1.9.0",
    "mysql-connector-python>=9.3.0",
    "paramiko>=3.5.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
